chore(util): move and fine tune InputMethodUtils
